Fix command checking of echo in _gen_rollback_cfg
Fix generated pattern for command "copy running-config ..." in
_gen_rollback_cfg. The command has to be explicit in the full form
"copy running-configuration ..." to avoid breakage when the command
echo (which is full form in case of OS10 10.5.1.3) is compared with
the short form patern and fails the check.

The resulting exception was:
netmiko.ssh_exception.NetmikoTimeoutException: Timed-out reading
channel, data not available.