Fix error handling in merge operations
Merge operations can cause errors that are reported by various
messages containing string "Error".

The error handling program path must not call rollback() function
which is unimplemented and raises NotImplementedError exception
that is subsequently caught in Napalm code and fails to propagate
to the custom code or Napalm CLI. Therefore the cause of the merge
error and even the fact that the merge failed is lost.